Post test

+++ src/main/java/kbsc/kbsc/domain/user/Repository/UserRepository.java
@@ -4,5 +4,5 @@ import kbsc.kbsc.domain.user.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 //JpaRepository<Repository가 사용할 Entity, Entity에서 사용되는 PK>
-public interface UserRepository extends JpaRepository<Users, Integer> {
+public interface UserRepository extends JpaRepository<Users, Long> {
 }
